refactor: SignUpScreen 내용을 SignUpActivity로 통합
