+++ app/src/main/java/org/sopt/at/SignUpActivity.kt
@@ -4,29 +4,13 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.BorderStroke
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextField
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import org.sopt.at.ui.theme.ATSOPTANDROIDTheme
 
 class SignUpActivity : ComponentActivity() {
@@ -39,4 +23,22 @@ class SignUpActivity : ComponentActivity() {
             }
         }
     }
+}
+
+@Preview(showBackground = true)
+@Composable
+fun SignUpScreen(modifier: Modifier = Modifier) {
+    var currentStep by remember { mutableStateOf(1) }
+    var id by remember { mutableStateOf("") }
+    var pw by remember { mutableStateOf("") }
+
+    when (currentStep) {
+        1 -> IdInputScreen(
+            id, { id = it },
+            onNext = { currentStep = 2 },
+        )
+
+        else -> PwInputScreen(pw, { pw = it })
+    }
+
 }
