--- app/src/main/java/org/sopt/at/SignUpScreen.kt
@@ -1,27 +1,0 @@
-package org.sopt.at
-
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-
-@Preview(showBackground = true)
-@Composable
-fun SignUpScreen(modifier: Modifier = Modifier) {
-    var currentStep by remember { mutableStateOf(1) }
-    var id by remember { mutableStateOf("") }
-    var pw by remember { mutableStateOf("") }
-
-    when (currentStep) {
-        1 -> IdInputScreen(
-            id, { id = it },
-            onNext = { currentStep = 2 },
-        )
-
-        else -> PwInputScreen(pw, { pw = it })
-    }
-
-}
